requirements.txt
boto3==1.35.39
chromadb==0.5.13 
flask[async]==3.0.3
gunicorn==23.0.0
JSON-log-formatter==1.1
markdown==3.7
piccolo[postgres]==1.22.0
psycopg2==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
python-slugify==8.0.4 
ollama==0.3.3 
requests==2.32.3 